+++ src/Vsix/TextViewCreationListener.cs
@@ -2,6 +2,7 @@
 using Microsoft.VisualStudio.Language.Intellisense;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Text.Operations;
 using Microsoft.VisualStudio.TextManager.Interop;
 using Microsoft.VisualStudio.Utilities;
 using System;
@@ -23,6 +24,9 @@ namespace EditorConfig
         [Import]
         internal SVsServiceProvider ServiceProvider = null;
 
+        [Import(typeof(ITextBufferUndoManagerProvider))]
+        private ITextBufferUndoManagerProvider UndoProvider { get; set; }
+
         private ErrorListProvider _errorList;
 
         public void VsTextViewCreated(IVsTextView textViewAdapter)
@@ -35,9 +39,15 @@ namespace EditorConfig
             if (_errorList == null)
                 return;
 
-            CommandFilter filter = new CommandFilter(view, CompletionBroker);
-            textViewAdapter.AddCommandFilter(filter, out var next);
-            filter.Next = next;
+            var filter = new CompletionController(view, CompletionBroker);
+            textViewAdapter.AddCommandFilter(filter, out var completionNext);
+            filter.Next = completionNext;
+
+            var undoManager = UndoProvider.GetTextBufferUndoManager(view.TextBuffer);
+
+            var formatter = new EditorConfigFormatter(view, undoManager);
+            textViewAdapter.AddCommandFilter(formatter, out var formatterNext);
+            formatter.Next = formatterNext;
 
             view.Closed += OnViewClosed;
         }