+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 - [x] Brace matching
 - [x] Drag 'n drop
 - [x] File icon for .editorconfig files
+- [x] Formatting command
 - [ ] Implement custom parser
-- [ ] Formatting command
 - [ ] Validate numeric values
 
 Features that have a checkmark are complete and available for